requirements.txt
Django==1.10.4
django-debug-toolbar==1.6
djangorestframework==3.5.3
django-crispy-forms==1.6.1
django-filter==1.0.1
django-multiselectfield==0.1.4

pytz==2016.10
pillow==3.4.2
redis==2.10.5
sorl-thumbnail==12.3
phonenumbers==7.7.5

mysqlclient==1.3.9
gunicorn==19.6.0
fabric3